requirements.txt
Django==5.2.12
django-ninja==1.6.0
dj-database-url==3.0.1
psycopg2-binary==2.9.9
python-decouple==3.8
requests==2.31.0
urllib3==2.2.1
celery==5.3.6
redis==5.0.3
django-unfold==0.45.0
gunicorn==21.2.0
uvicorn==0.27.1
boto3==1.34.11
django-storages==1.14.2
